# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Newsight API
  description: REST API documentation for the Newsight backend.
  contact:
    name: Newsight
  version: v1
servers:
- url: http://localhost:8080
  description: Local backend
- url: https://newsightkr.com
  description: Production
...
paths:
...
  /api/inquiries:
    get:
      tags:
      - User Inquiries
      summary: List inquiries
      operationId: listInquiries
      parameters:
      - name: inquiry_type
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: mine
        in: query
        required: false
        schema:
          type: boolean
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InquiryListResponseDto"
      security:
      - bearerAuth: []
    post:
      tags:
      - User Inquiries
      summary: Create inquiry
      operationId: createInquiry
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InquiryCreateRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InquiryCreateResponseDto"
      security:
      - bearerAuth: []
  /api/inquiries/:
    get:
      tags:
      - User Inquiries
      summary: List inquiries
      operationId: listInquiries_1
      parameters:
      - name: inquiry_type
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: mine
        in: query
        required: false
        schema:
          type: boolean
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InquiryListResponseDto"
      security:
      - bearerAuth: []
    post:
      tags:
      - User Inquiries
      summary: Create inquiry
      operationId: createInquiry_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InquiryCreateRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InquiryCreateResponseDto"
      security:
      - bearerAuth: []
...
  /api/inquiries/{inquiry_seq}/:
    get:
      tags:
      - User Inquiries
      summary: Get inquiry detail
      operationId: getInquiryDetail
      parameters:
      - name: inquiry_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InquiryDetailResponseDto"
      security:
      - bearerAuth: []
  /api/inquiries/{inquiry_seq}:
    get:
      tags:
      - User Inquiries
      summary: Get inquiry detail
      operationId: getInquiryDetail_1
      parameters:
      - name: inquiry_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InquiryDetailResponseDto"
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    InquiryCreateRequestDto:
      type: object
      properties:
        inquiry_type:
          type: string
          minLength: 1
        title:
          type: string
          maxLength: 200
          minLength: 0
        message:
          type: string
          minLength: 1
        is_private:
          type: boolean
      required:
      - inquiry_type
      - message
      - title
    InquiryCreateResponseDto:
      type: object
      properties:
        inquiry_seq:
          type: integer
          format: int64
        submitted_at:
          type: string
          format: date-time
...
    InquiryDetailResponseDto:
      type: object
      properties:
        inquiry:
          $ref: "#/components/schemas/InquiryDto"
    InquiryDto:
      type: object
      properties:
        inquiry_seq:
          type: integer
          format: int64
        inquiry_type:
          type: string
        title:
          type: string
        content:
          type: string
        writer_user_id:
          type: string
        created_at:
          type: string
          format: date-time
        status:
          type: string
        is_private:
          type: boolean
        admin_message:
          type: string
        processed_at:
          type: string
          format: date-time
        answer_updated_at:
          type: string
          format: date-time
        answered_by:
          type: string
        answer_team_label:
          type: string
    InquiryItemDto:
      type: object
      properties:
        inquiry_seq:
          type: integer
          format: int64
        inquiry_type:
          type: string
        title:
          type: string
        writer_user_id:
          type: string
        created_at:
          type: string
          format: date-time
        status:
          type: string
        is_private:
          type: boolean
    InquiryListResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/InquiryItemDto"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        total_count:
          type: integer
          format: int64
        total_pages:
          type: integer
          format: int32
...
  securitySchemes:
    bearerAuth:
      type: http
      description: Use the access token returned by the login API.
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT